+++ codegen/src/commonMain/kotlin/com/freeletics/khonshu/codegen/NavHostActivity.kt
@@ -25,6 +25,7 @@ public annotation class NavHostActivity(
     val parentScope: KClass<*> = AppScope::class,
     val stateMachine: KClass<out StateMachine<*, *>>,
     val activityBaseClass: KClass<*>,
+    val experimentalNavigation: Boolean = false,
 )
 
 /**